Hide `patterns` module and structs from users
- Rename directory `patterns` to `pattern`.
- Simplify the comments of pattern structs.
