+++ spdlog/src/formatter/pattern_formatter/pattern/level.rs
@@ -7,9 +7,6 @@ use crate::{
 
 /// A pattern that writes the level of a log record into the output. Examples:
 /// `critical`, `error`, `warn`.
-///
-/// This pattern corresponds to `{l}` or `{level}` in the pattern template
-/// string.
 #[derive(Copy, Clone, Debug, Default)]
 pub struct Level;
 
@@ -35,9 +32,6 @@ impl Pattern for Level {
 
 /// A pattern that writes the level in a shorter form of a log record into the
 /// output. Examples: `C`, `E`, `W`.
-///
-/// This pattern corresponds to `{L}` or `{level-short}` in the pattern template
-/// string.
 #[derive(Copy, Clone, Debug, Default)]
 pub struct ShortLevel;
