+++ spdlog/src/formatter/pattern_formatter/pattern/mod.rs
